+++ demo/wrapped-components/RadioGroup.js
@@ -2,8 +2,8 @@ import React, { forwardRef, useImperativeHandle, useRef } from "react";
 import {
   useAttribute,
   useBooleanAttribute,
-  useProperties,
   useEventListener,
+  useProperties,
 } from "./react-utils.js";
 import "../components/radio-group.js";